Moved files around a bit.  Might make a new folder per game in src/ eventually.
