+++ src/geometry.c
