+++ windows/Player/Player/Player.vcxproj
@@ -64,15 +64,15 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\..\arkanoid.c" />
-    <ClCompile Include="..\..\..\geometry.c" />
+    <ClCompile Include="..\..\..\src\arkanoid.c" />
+    <ClCompile Include="..\..\..\src\geometry.c" />
     <ClCompile Include="screen.cpp" />
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\..\arkanoid.h" />
-    <ClInclude Include="..\..\..\game.h" />
-    <ClInclude Include="..\..\..\geometry.h" />
+    <ClInclude Include="..\..\..\include\arkanoid.h" />
+    <ClInclude Include="..\..\..\include\game.h" />
+    <ClInclude Include="..\..\..\include\geometry.h" />
     <ClInclude Include="screen.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
